criação de contexto para salvar o login do usuário em qualquer rota
