+++ src/pages/NewRoom.tsx
@@ -1,16 +1,15 @@
-import { useContext } from 'react';
 import { Link } from 'react-router-dom';
-import { AuthContext } from '../App';
 
 import illustrationImg from '../assets/images/illustration.svg'
 import logoImg from '../assets/images/logo.svg';
 
 import { Button } from '../components/Button';
+import { useAuth } from '../hooks/useAuth';
 
 import '../styles/auth.scss';
 
 export function NewRoom() {
-    const { user } = useContext(AuthContext)
+    const { user } = useAuth()
 
 
     return (
@@ -23,7 +22,6 @@ export function NewRoom() {
             <main>
                 <div className="main-content">
                     <img src={logoImg} alt="Letmeask" />
-                    <h1>Olá, {user?.name} <br></br> como você está hoje?</h1>
                     <h2>Criar uma nova sala</h2>
                     <form>
                     <input type="text"
